pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deadlock-graph-tools"
version = "0.1.0"
description = "Tooling and datasets for maintaining a Deadlock character matchup graph in Neo4j."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "neo4j>=5.20",
    "pydantic>=2.7",
    "pydantic-settings>=2.7",
    "typer>=0.12",
    "python-dotenv>=1.0",
    "pyyaml>=6.0"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]

[tool.setuptools.packages.find]
where = ["src"]
